Backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
platformdirs==4.4.0
PyJWT==2.10.1
virtualenv==20.34.0
Werkzeug==3.1.3
gunicorn==23.0.0

